tests now run
